pulled db interactors out of index route
